Added typing to arguements

+++ src/calculator.py
@@ -1,12 +1,12 @@
 class Calculator:
     def add(self, x:int, y:int)->int:
-        return x + y
+        return int(x + y)
 
     def subtract(self, x:int, y:int)->int:
-        return x - y
+        return int(x - y)
 
     def multiply(self, x:int, y:int)->int:
-        return x * y
+        return int(x * y)
 
     def divide(self, x:int, y:int)->int:
         if y == 0:
